Extract relation-related functions
This also moves relation-related tests into common (from ingest, where
they didn't make much sense).
