+++ src/common/src/test/resources/relation-191204.wkt
