+++ src/common/src/test/resources/relation-2554903.wkt
