+++ src/common/src/test/resources/relation-2580685.wkt
