+++ src/common/src/test/resources/relation-3105056.wkt
