+++ src/common/src/test/resources/relation-333501.wkt
