+++ src/common/src/test/resources/relation-393502.wkt
