+++ src/common/src/test/resources/relation-5448156.wkt
