+++ src/common/src/test/resources/relation-5612959.wkt
