+++ src/common/src/test/resources/relation-6710544.wkt
